+++ Tests/VCoreTests/Tests/Extensions/Foundation/Collection/CollectionNonEmptyTests.swift
@@ -1,0 +1,17 @@
+//
+//  CollectionNonEmptyTests.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 17.11.23.
+//
+
+import XCTest
+@testable import VCore
+
+// MARK: - Tests
+final class CollectionNonEmptyTests: XCTestCase {
+    func test() {
+        XCTAssertNil(Array<Int>().nonEmpty)
+        XCTAssertNotNil([1, 2, 3].nonEmpty)
+    }
+}