+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Views
 
 Extensions
 
+- `Collection.nonEmpty` property is added, that returns non-empty `Collection`, or `nil`.
 - `withContinuousObservationTracking(of:on:execute:)` method is added, that tracks access to properties in a `class`-type continuously
 
 Extra - XCode Templates